adds a totals module
