+++ lib/modules/sales_analyst_helper_methods_module.rb
@@ -1,20 +1,5 @@
 module SAHelper
 
-  def total_merchants
-    @sales_engine.merchants.storage.length
-  end
-
-  def total_items
-    @sales_engine.items.storage.length
-  end
-
-  def total_items_per_merchant
-    @sales_engine.items.storage.reduce(Hash.new(0)) do |hash, item|
-      hash[item.merchant_id] += 1
-      hash
-    end
-  end
-
   def merchant_id_with_high_item_count
     total_items_per_merchant.map do |merch, item|
       if item > (average_items_per_merchant + average_items_per_merchant_standard_deviation)
@@ -60,20 +45,8 @@ module SAHelper
     standard_deviation(average, collection)
   end
 
-
-  def total_invoices
-    @sales_engine.invoices.all
-  end
-
   def two_standard_deviations
     average_invoices_per_merchant_standard_deviation * 2
-  end
-
-  def total_invoices_per_merchant
-    @sales_engine.invoices.storage.reduce(Hash.new(0)) do |hash, item|
-      hash[item.merchant_id] += 1
-      hash
-    end
   end
 
   def top_merchants_by_invoice_count_merchant_ids
@@ -85,15 +58,6 @@ module SAHelper
   def bottom_merchants_by_invoice_count_merchant_ids
     total_invoices_per_merchant.keys.find_all do |merchant_id|
       total_invoices_per_merchant[merchant_id] < (average_invoices_per_merchant - two_standard_deviations)
-    end
-  end
-
-
-  def total_invoices_per_day_hash
-    @sales_engine.invoices.storage.reduce(Hash.new(0)) do |hash, item|
-      time = item.created_at.strftime("%A")
-      hash[time] += 1
-      hash
     end
   end
